do not re-render styles if variant override provided
