+++ libs/designer/src/components/Editor/EditorPanels/RightSidebar/StylePanel/Declarations/NewDeclaration.tsx
