+++ libs/designer/src/components/Editor/EditorPanels/RightSidebar/StylePanel/Declarations/cssSchema.ts
