+++ libs/designer/src/styles/theme.pc
@@ -49,6 +49,8 @@ public token roundedSM 0.125rem
 public token rounded 0.25rem
 public token roundedMD 0.375rem
 
+public token inputHeight 24px
+
 public style mask {
     background-color: currentColor
     -webkit-mask-size: 100%
@@ -77,7 +79,7 @@ public style input {
     border-radius: var(rounded)
     border: none
     color: currentColor
-    min-height: 24px
+    min-height: var(inputHeight)
 }
 
 public style card {
